requirements.txt
chainlit
langchain
langchain-groq
langchain-postgres
langchain-huggingface
langchain-experimental
langchain-community
langchain-classic
langgraph
pandas
asyncpg
python-dotenv
sentence-transformers
psycopg[binary,pool]
aiosqlite